Add and integrate new PlacableActors
Finishes Chapter 3 Challenge.

Adds `ShadowActivator` and `ShadowPlayer` placeable actors to the project. When the player steps on a `ShadowActivator`, the `ShadowPlayer` will be activated. This shadow player will mirror the player's movements, and will be able to collect map items, unlock doors, but also take damage that affects the player.