Rework virtual functions to reduce copied code
Setting the color of the placeable actor before drawing it was a common bit of code that could really be moved into the Draw() class. The only thing that changed for objects was the state of their color- I made GetColor virtual so that a class could update its color based on the state.

There is now less duplicated code, and the expected behavior of writing a single character to the screen is supported in the base class (although still marked virtual for any other classes that would want to completely override the behavior).