Module 3 Chapter 2 Challenge Pt2
Refactored `GameplayState::HandleCollision` to use a new class `CollisionEngine`. This removes most of the responsibility of collision management from `GameplayState`. It could be refactored further to move all collision management into this class, and could avoid type checking using the visitor pattern, but those are outside the scope of this current assignment.