Module 4 Chapter 2 preserve temp chat message
Previously, when a chat participant received a message from another participant, if they had a message in progress on the command line, it would be lost. Now, there is a separate thread in charge of buffering the user's input so that the message is not lost when they receive a new message.